+++ log/elasticsearch/elasticsearch.go
@@ -13,18 +13,21 @@ import (
 )
 
 import (
+	"encoding/json"
 	"github.com/pkg/errors"
 	es "gopkg.in/olivere/elastic.v3"
 )
 
 type EsClient struct {
 	*es.Client
+	bulk *es.BulkService
 }
 
 func CreateEsClient(hosts []string) (EsClient, error) {
 	var esClient EsClient
 	// Create a client
-	client, err := es.NewClient(es.SetURL(hosts[:]...))
+	// snif的作用是根据部分url获取整个集群的urls
+	client, err := es.NewClient(es.SetURL(hosts[:]...), es.SetSniff(true), es.SetHealthcheck(true))
 	if err == nil {
 		esClient.Client = client
 	}
@@ -98,6 +101,7 @@ func (ec EsClient) Insert(index string, typ string, msg interface{}) error {
 	// https://github.com/olivere/elastic/issues/127
 	// Elasticsearch can create an identifier for you, automatically.
 	// _, err = ec.Index().Index(index).Type(typ).Id(1).BodyJson(msg).Do()
+	ctx = context.Background()
 	switch msg.(type) {
 	case string:
 		_, err = ec.Index().Index(index).Type(typ).BodyString(msg.(string)).DoC(ctx)
@@ -122,9 +126,6 @@ func (ec EsClient) Insert(index string, typ string, msg interface{}) error {
 	return nil
 }
 
-// Search
-// Waiting for es5.x's future sql feature
-
 // InsertWithDocId 插入@msg时候指定@docID
 // !!! 如果@msg的类型是string 或者 []byte，则被当做Json String类型直接存进去
 func (ec EsClient) InsertWithDocId(index string, typ string, docID string, msg interface{}) error {
@@ -135,6 +136,7 @@ func (ec EsClient) InsertWithDocId(index string, typ string, docID string, msg i
 		ctx      context.Context
 	)
 
+	ctx = context.Background()
 	switch msg.(type) {
 	case string:
 		_, err = ec.Index().Index(index).Type(typ).Id(docID).BodyString(msg.(string)).DoC(ctx)
@@ -158,3 +160,53 @@ func (ec EsClient) InsertWithDocId(index string, typ string, docID string, msg i
 
 	return nil
 }
+
+// BulkInsert 批量插入@arr
+// !!! 如果@arr[0]的类型是string 或者 []byte，则被当做Json String类型直接存进去
+// https://www.elastic.co/guide/en/elasticsearch/guide/current/bulk.html
+// A good place to start is with batches of 1,000 to 5,000 documents
+func (ec EsClient) BulkInsert(index string, typ string, arr []interface{}) error {
+	var (
+		err  error
+		ctx  context.Context
+		bulk *es.BulkService
+		rsp  *es.BulkResponse
+	)
+
+	if ec.bulk == nil {
+		ec.bulk = ec.Bulk()
+	}
+
+	bulk = ec.bulk.Index(index).Type(typ)
+	for _, e := range arr {
+		switch e.(type) {
+		case string:
+			data := ([]byte)(e.(string))
+			bulk.Add(es.NewBulkIndexRequest().Doc((*json.RawMessage)(&data)))
+		default:
+			if data, ok := e.([]byte); ok {
+				bulk.Add(es.NewBulkIndexRequest().Doc((*json.RawMessage)(&data)))
+			} else {
+				bulk.Add(es.NewBulkIndexRequest().Doc(e))
+				// bulk.Add(es.NewBulkIndexRequest().Index("1").Doc(e))
+			}
+		}
+	}
+	if bulk.NumberOfActions() <= 0 {
+		return fmt.Errorf("bulk.NumberOfActions() = %d", bulk.NumberOfActions())
+	}
+
+	ctx = context.Background()
+	rsp, err = bulk.DoC(ctx)
+	if err != nil {
+		return errors.Wrapf(err, "BulkInsert(@arr:%#v)", arr)
+	}
+	if rsp.Errors {
+		return fmt.Errorf("BulkInsert(@arr:%#v) = error:%#v ", arr, rsp.Failed())
+	}
+
+	return nil
+}
+
+// Search
+// Waiting for es5.x's future sql feature