+++ log/elasticsearch/elasticsearch_test.go
@@ -2,6 +2,7 @@ package gxelasticsearch
 
 import (
 	"encoding/json"
+	"strconv"
 	"testing"
 )
 
@@ -174,6 +175,82 @@ func TestEsClient_InsertWithString(t *testing.T) {
 	}
 	docString, _ = json.Marshal(doc)
 	err = client.InsertWithDocId(esConf.PushIndex, esConf.PushType, key, string(docString))
+	if err != nil {
+		t.Fatal(err)
+	}
+}
+
+// go test -v -run BulkInsert
+func TestEsClient_BulkInsert(t *testing.T) {
+	var (
+		err      error
+		esConf   EsConf
+		client   EsClient
+		docArray []interface{}
+		doc      Doc
+		docBytes []byte
+	)
+
+	esConf = EsConf{
+		ShardNum:        5,
+		ReplicaNum:      0,
+		RefreshInterval: 1,
+		EsHosts: []string{
+			"http://119.81.218.90:5858",
+		},
+		PushIndex: "dokidoki-push-test",
+		PushType:  "gopush",
+	}
+
+	// Create a client
+	client, err = CreateEsClient(esConf.EsHosts)
+	if err != nil {
+		t.Fatal(err)
+	}
+
+	// Delete the index again
+	err = client.DeleteEsIndex(esConf.PushIndex)
+	if err != nil {
+		t.Log(err)
+	}
+
+	// Create an index
+	err = client.CreateEsIndex(esConf.PushIndex, esConf.ShardNum, esConf.ReplicaNum, esConf.RefreshInterval)
+	if err != nil {
+		t.Fatal(err)
+	}
+
+	// Add Doc Array
+	for i := 0; i < 10; i++ {
+		docArray = append(docArray, Doc{
+			Subject: "Invitation index:" + strconv.Itoa(i),
+			Message: "Would you like to visit me in Berlin in October?",
+		})
+	}
+
+	err = client.BulkInsert(esConf.PushIndex, esConf.PushType, docArray)
+	if err != nil {
+		t.Fatal(err)
+	}
+
+	// Delete the index again
+	err = client.DeleteEsIndex(esConf.PushIndex)
+	if err != nil {
+		t.Log(err)
+	}
+
+	docArray = docArray[:0]
+	// Add Doc String Array
+	for i := 0; i < 10; i++ {
+		doc = Doc{
+			Subject: "Invitation index:" + strconv.Itoa(i*10),
+			Message: "Would you like to visit me in Berlin in October?",
+		}
+		docBytes, _ = json.Marshal(doc)
+		docArray = append(docArray, docBytes)
+	}
+
+	err = client.BulkInsert(esConf.PushIndex, esConf.PushType, docArray)
 	if err != nil {
 		t.Fatal(err)
 	}